Reset style source when instance is selected
